+++ cmd/firewall_delete.go
@@ -1,6 +1,8 @@
 package cmd
 
 import (
+	"fmt"
+
 	"github.com/exoscale/egoscale"
 
 	"github.com/spf13/cobra"
@@ -15,6 +17,18 @@ var firewallDeleteCmd = &cobra.Command{
 		if len(args) < 1 {
 			return cmd.Usage()
 		}
+
+		force, err := cmd.Flags().GetBool("force")
+		if err != nil {
+			return err
+		}
+
+		if !force {
+			if !askQuestion(fmt.Sprintf("sure you want to delete %q security group", args[0])) {
+				return nil
+			}
+		}
+
 		return deleteFirewall(args[0])
 	},
 }
@@ -33,5 +47,6 @@ func deleteFirewall(name string) error {
 }
 
 func init() {
+	firewallDeleteCmd.Flags().BoolP("force", "f", false, "Attempt to remove security group without prompting for confirmation")
 	firewallCmd.AddCommand(firewallDeleteCmd)
 }
